% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterabilitytest.R
\name{print.clusterability}
\alias{print.clusterability}
\title{Print a clusterability object}
\usage{
\method{print}{clusterability}(x, ...)
}
\arguments{
\item{x}{An object of class 'clusterability'}

\item{...}{Not used}
}
\description{
Print function to display results from a clusterability test.
}
\seealso{
\code{\link{clusterabilitytest}}
}
